Refactor User Interface selectizeInput Function

This function aim to simplify the repetitive call of shiny's selectizeInput function. It will be called in app.R file.

ui_selectize(id, choices)

Arguments

id
Input id that will be used in the server side.

choices
List of value to select.

Examples

ui_selectize("country_select", c("Italy", "Spain", "France", "Austria", "United Kingdom"))

#> <div class="form-group shiny-input-container">
#>   <label class="control-label" for="country_select">Step 1: Select or type at least 1 country</label>
#>   <div>
#>     <select id="country_select" class="form-control" multiple="multiple"><option value="Italy" selected>Italy</option>
#> <option value="Spain" selected>Spain</option>
#> <option value="France">France</option>
#> <option value="Austria" selected>Austria</option>
#> <option value="United Kingdom">United Kingdom</option></select>
#>     <script type="application/json" data-for="country_select">{}</script>
#>   </div>
#> </div>